SX: improve color detection to support rgba() and hsla()
This commit introduces and improved detection of colors. Specifically, it understands hexadecimal RGB syntaxes with alpha channel as well as rgb(), rgba(), hsl() and hsla(). The functional syntaxes are simplified (no whitespace syntax support, weak floats values, no further validation like mixing percentages and numbers). This can be improved later when needed.

We need to improve this `isColor` function because some parts of the SX system rely on this detection (for example expansion of shorthand properties where we need to detect what is a color in `background` property).

I am also considering a new package `@adeira/css-color` because it might be quite handly when we master this.

adeira-source-id: 28be2533e8d80cce8226d5f5d26d98eba2402ae0