sx: Calculate insertIndex inside loop in injectRuntimeStyles
This was defined outside the loop before, it was ok but confusing.
It is easy to think that this works like an array and that you are
overwriting index, though it is just adding and pushing the other down.
Also added a test for that the runtime styles are added correctly

adeira-source-id: 0b35504deff38645493fb7cc2fe39fc92ce24afa