SX: improve developer experience
Previously, it was possible to define an empty stylesheet definition,
however, it was not possible to call it which kind of defeated the
purpose. I noticed that it's quite handy to be able to define
empty stylesheet from time to time (when developing) and this was an
annoying limitation. Fixed now.

adeira-source-id: d7e1f312801147626bda880f9f72da1a96a60937